Updated core to accept new argument.
The core now takes a -M and a package_name for mirroring a package.
There are some changes on how root is now set. It now checks to see if
the update_directory is already in roots, if it is it will use that. If
not it will accept the default, or will grab the first item from roots.

Should probably also consider moving away from getopt and move to
optparse or argparse for better argument handling and expanding options
for pypiserver.