Drop support for Python 3.4 (#321)
* Drop support for Python 3.4

Python 3.4 is no longer supported. Even pyenv is failing to install it
for me, because apparently the currnet version of `ensurepip` bombs for
3.4. Pypiserver may still work on 3.4, but testing on it has becomes
more of a hassle than it's worth.

* Fix @mplanchard's email address

Just realized my email address in the authors' file has been wrong for
pretty much forever.

* Remove GL CI

GitLab CI is nice, but doesn't support forks, and so isn't going to be
viable for this project.