Add PypiserverApp app factory
This adds a `PypiserverApp` app factory to make the process of
instantiating the app, providing a config, etc. more straightforward
and easy to reason about. The factory is instantiated with a config
and a Bottle instance, and the `_initialize_app()` method adds routes
and hooks. Methods on the class have access to the config as needed.

This avoids the need to mutate the top-level global scope of the `_app`
module, along with the need to pop it out of `sys.modules`. While
the end result is the same, hopefully this is a bit easier for newcomers
to reason about.

Since we no longer need to do odd things with the import path, the
`_app` module has been renamed to `pypiserver_app`.