Fixed violation of PEP 503 warning with pip>=22
When using pip >=22, we get the following warning:

"DEPRECATION: The HTML index page being used (https://....) is not a proper
HTML 5 document. This is in violation of PEP 503 which requires these pages
to be well-formed HTML 5 documents. Please reach out to the owners of this
index page, and ask them to update this index page to a valid HTML 5
document. pip 22.2 will enforce this behaviour change. Discussion can be
found at https://github.com/pypa/pip/issues/10825"