Deduplicate host.
pypiserver will use `X-Forwarded-Host` to determine the Host
for redirects. Because of an odity with nginx ingress this
headers is being set twice, and it is not trivially for us to
remove the first. Instead we have patched pypiserver to detected
when multiple host entries are returned, and only to take the last
one.